requirements.txt
cobs==1.1.4
scipy==1.5.2
pyqtgraph==0.11.0rc0
psutil==5.7.3
matplotlib==3.3.1
numpy==1.19.1
sounddevice==0.3.15
dataclasses==0.8
PyQt5==5.15.2
pyserial==3.5
